feat: added physical memory allocator still not checking for invalid memory areas

+++ src/multiboot.h
@@ -38,8 +38,8 @@ typedef struct MULTIBOOT_HEADER_STRUCT mboot_header_T;
 struct MULTIBOOT_MEMORY_MAP_STRUCT
 {
     uint32_t size;
-    uint32_t base_addr;
-    uint32_t length;
+    uint64_t base_addr;
+    uint64_t length;
     uint32_t type;
 
 } __attribute__((packed));
